pyproject.toml
[project]
name = "vertexservermanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil",
    "requests",
    "pyyaml",
    "schedule",
]
